feat: added context to authReducer
