+++ src/reducer/authReducer.js
@@ -1,0 +1,13 @@
+export const authReducer = (state, action) => {
+  switch (action.type) {
+    case "AUTH_SUCCESS":
+      return {
+        ...state,
+        isAuth: action.payload.encodedToken ? true : false,
+        user: action.payload.user,
+        encodedToken: action.payload.encodedToken,
+      };
+    default:
+      return state;
+  }
+};
