+++ src/reducer/index.js
@@ -1,1 +1,2 @@
-export {productsFilterReducer} from "./productsFilterReducer"
+export { productsFilterReducer } from "./productsFilterReducer";
+export { authReducer } from "./authReducer";
